+++ Commands/Properties/AssemblyInfo.cs
@@ -48,6 +48,6 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("3.20.2004.0")]
-[assembly: AssemblyFileVersion("3.20.2004.0")]
+[assembly: AssemblyVersion("3.21.2005.0")]
+[assembly: AssemblyFileVersion("3.21.2005.0")]
 [assembly: InternalsVisibleTo("SharePointPnP.PowerShell.Tests")]