+++ CHANGELOG.md
@@ -10,8 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 ### Changed
+- Fixed uploading a file using `Add-PnPFile` using `-ContentType` throwing an exception [PR #2619](https://github.com/pnp/PnP-PowerShell/pull/2619)
 
 ### Contributors
+- Eduardo Garcia-Prieto [egarcia74]
 
 ## [3.20.2004.0]