+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 ### Contributors
+- Rune Sperre [rsperre]
 
 ## [3.20.2004.0]